reworking queue (continued): merged url queue into main download queue: urls added to queue are now immediately shown in web-interface; urls can be reordered and deleted; when urls are fetched the downloaded nzb-files are put into queue at the positions of their urls; this solves the problem with fetched nzb-files ordered differently than the urls if the fetching of upper (position wise) urls were completed after of the lower urls; removed options "ReloadUrlQueue" and "ReloadPostQueue" since there are no separate url- and post-queues anymore; nzb-files added via urls have new field "URL" which can be accessed via RPC-methods "listgroups" and "history"; new env. var. "NZBNP_URL", "NZBNA_URL" and "NZBPP_URL" passed to NzbProcess, NzbAddedProcess and PostProcess-scripts; removed remote command "--list U", urls are now shown as groups by command "--list G"; RPC-method "urlqueue" is still supported for compatibility but should not be used since the urls are now returned by method "listgroups", the entries have new field "Kind" which can be "NZB" or "URL"

+++ trunk/daemon/postprocess/ParChecker.cpp
@@ -770,7 +770,7 @@ bool ParChecker::AddMissingFiles()
 
 			int iWasMissing = ((Repairer*)m_pRepairer)->missingfilecount;
 			((Repairer*)m_pRepairer)->VerifyExtraFiles(extrafiles1);
-			bool bAdded = iWasMissing > ((Repairer*)m_pRepairer)->missingfilecount;
+			bool bAdded = iWasMissing > (int)((Repairer*)m_pRepairer)->missingfilecount;
 			if (bAdded)
 			{
 				info("Found missing file %s", Util::BaseFileName(pExtraFile->FileName().c_str()));
